fix: fix conditional statement in merge.yaml
The conditional statement in merge.yaml was not correctly checking for the user login of dependabot[bot] or pre-commit-ci[bot]. This has been fixed by using a multi-line conditional statement in the if block.

feat: Rename workflow name in template.yaml

The name of the workflow in template.yaml has been renamed from "Sync with Template Repository" to "Template Repository" for clarity and consistency.

fix: Fix command in init.mk

The command in init.mk for installing a file has been fixed to remove the unnecessary "-" before the "install" command. This ensures that the file is properly installed with the correct permissions. Additionally, the setup-github.sh script is now executed with the "- " prefix to handle any errors that may occur.

Why the changes were done:
- The conditional statement in merge.yaml was fixed to ensure that the correct users are targeted for auto-merge.
- The workflow name in template.yaml was renamed to provide a more concise and descriptive name.
- The command in init.mk was fixed to properly install files and handle any errors that may occur during the execution of the setup-github.sh script.