feat(action): add support for GitHub token
This commit adds a new input `token` to the action.yaml file. The `token` input allows users to provide their GitHub token, which can be used for authentication and authorization when interacting with the GitHub API.

The `token` input is optional and has a default value of `${{ github.token }}`, which is the token automatically provided by GitHub Actions.

The added environment variable `GH_TOKEN` is used to pass the `token` input value to the subsequent steps in the action workflow.