fix: add ignorePaths for lock files
Add `ignorePaths` configuration to the cspell.config.yaml file to exclude lock files from spell checking. This is done to prevent false positives and improve the accuracy of the spell checker.

Also, update the cspell-init.py script to exclude lock files when generating the list of words for spell checking. This ensures that lock files are not included in the spell check process.

The lock files are now ignored in both the configuration file and the script.