fix(workflows): fix condition for merge job
The condition for the merge job in the `.github/workflows/merge.yaml` file was incorrect. It was checking for the login of the pull request user to be either 'dependabot[bot]' or 'pre-commit-ci[bot]'. However, the condition was not properly formatted and was causing the job to not run as expected. This commit fixes the condition by removing the unnecessary `|` and properly formatting the condition.

This commit addresses the issue by updating the condition for the merge job in the `.github/workflows/merge.yaml` file.