feat(dependabot): add groups for different package ecosystems
This commit adds groups for different package ecosystems in the Dependabot configuration file.
The groups allow for more granular control over which packages are updated in each ecosystem.
This change improves the organization and management of dependencies in the project.

feat(workflows): remove check job from merge workflow

This commit removes the check job from the merge workflow.
The check job was responsible for waiting for all checks to complete before merging the pull request.
However, this functionality is no longer needed as it is handled by GitHub Actions automatically.
Removing the check job simplifies the workflow and reduces unnecessary complexity.