fix: update Dependabot schedule to daily
Update the Dependabot schedule for package ecosystems (cargo, gitsubmodule, github-actions, gomod, npm, pip) to run daily instead of weekly. This change ensures that dependencies are checked and updated more frequently, improving the overall security and stability of the project.

Add permissions for writing contents in the Merge workflow to allow for smoother integration and automation processes.

Remove the scheduled cron job from the Template Repository workflow as it is no longer needed. Update permissions to allow writing contents in the Template workflow for better management.