fix: add caching for pre-commit to improve workflow efficiency
To optimize the workflow efficiency, a caching mechanism has been added for pre-commit. This will cache the pre-commit setup in the ~/.cache/pre-commit directory, using the hash of the .pre-commit-config.yaml file as the key. This change ensures that the pre-commit setup is cached and reused, reducing setup time for subsequent runs.