fix(workflows): add permissions for actions in template.yaml
Added permissions for actions in the template job in the template.yaml file to allow writing actions. Also, added a token input with the GitHub token in the steps. This change ensures that the actions have the necessary permissions to write in the template job.

fix(action): update action.yaml to use inputs.commit

Updated the action.yaml file to use the `inputs.commit` instead of `push` for committing changes. Removed the deprecated `push` input and made the necessary adjustments to ensure the correct usage of the `commit` input.

fix(init.mk): ignore errors in make command in init.mk

Modified the make command in the init.mk file to include the `--ignore-errors` flag. This change allows the make command to continue execution even if errors occur, ensuring smoother synchronization with the repository template.